requirements.txt
# Web framework
Django==4.2.5
djangorestframework==3.14.0

# Pinecone
pinecone-client==2.2.1

# Transformers (Hugging Face for tokenization and model loading)
transformers==4.33.0

# PyTorch (for model inference)
torch==2.0.1

# OpenAI API (for ChatGPT integration)
openai==0.28.0

# Pandas (for handling Excel files)
pandas==2.0.3

# Other dependencies
requests==2.31.0  # Used by many libraries including Pinecone and OpenAI
